Add www files
These are what I found on the server, and they are copied directly from
the server. They should give us a starting point for refining the UI.
The next step is to set up a deploy process.

I left out the private files, such as passwords, and some other auxiliary
configuration files, such as the certbot config.