Make docker environment for web UI
This will make it easier for people to run and test the web UI component
without needing access to the production server. Using docker gives us a
way to test Apache configuration changes and will allow for us to
migrate the web to new technologies.

It doesn't handle the password access yet.