Get apache reverse proxy settings working
The python app is mounted under /p/. Need to pass SCRIPT_NAME header
so the python templating will generate correct links. Also need to get
Apache to send the prefix (it was being stripped), and need python app
to not redirect if a slash is missing (this caused an infinite redirect
loop).

Fixed function name error in the flask code.