Add flask-login for user management
This stores the user login info in a session cookie instead of having
the browser sent authoriztion headers. It still uses the Apache htaccess
file as the password database.

Using the session cookie to identify a log-in is on the path to adding
an oauth login.

Also remove the Apache rewriting of the paths. The flask script knows it
is run under a path-prefix, and apache was then rewriting the redirects
to add the path prefix a second time.