Add support for reading parton info from database
Many infrastructure changes:
* Set up database set-up and migration framework (yoyo migrations)
* Add `User` model
* Read user information from database after authentication
* Make user record if user missing from database

The authentication step (htaccess) identifies users by username. But
then the code uses the numeric rowid to identify users in the session.
Using the rowid in the session is because when the authentication is
changed, we won't have username as an identifier, will will have ORCID,
and so it seemed better to use the id number to identify people inside
the web app. The compute code still uses username.