Add queued jobs to the carrel database
Queued jobs have the status "queued".

With this status we can show the carrel in a user's carrel list before
the computation finishes processing. Having a carrel "disappear" between
being queued and being finished made it hard for users to reason about
how the system worked.

Change the carrel submittions to redirect to the "processing" carrel
page.