Batch carrel scraping into fewer sql updates
Instead of updating the carrel database after scanning each carrel,
batch the carrel updates into a fewer sqlite3 invocations. Due to a
limitation in sqlite3 we cannot update more than 500 carrels at a time.
(we were getting a "too many terms in compound SELECT" error when we
did).

This change is to improve performace and to try to limit the amount of
lock contention on the central carrel database file. It also speeds up
the carrel scraping.