Use random shortname if given something unusable
Sometimes secure_filename() will return an empty string. Also, since the
shortname is used to name a directory on disk, we want to make sure it
does not contain any characters that the command line tools might have
problems with.
So we first call secure_filename() which will try to convery unicode
characters into a ascii approximation. Then if that does not give us a
non-empty string containing only ascii letters and numerals, we generate
a random 8 character name and use that instead. Not perfect, but
prevents a job submission from failing.